+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 _general description_
 
-#### [0.4.2] job system (_23 jun 22_)
+#### [0.4.2] name (_23 jun 22_)
 
 - **added** - random number generator
 - **changed** - better concept names and organization
+- **added** - vector algebra
+- **todo** - global/local job queues
 
 #### [0.4.1] job system (_21 jun 22_)